fix(ci): coverage shard — extract redaction helpers to keep provider_readiness_helpers under 1500-line limit
The Ollama auto-pull work grew provider_readiness_helpers.py past the 1500-line
maintainability guardrail (test_first_party_code_files_stay_under_line_limit).
Move the self-contained secret-redaction/redacted-logging helpers into a new
provider_readiness_redaction module (a leaf concern depending only on the
redaction constants in provider_readiness) and re-export them so existing call
sites and tests keep reaching them via the original namespaces.
