+++ src/awf/service/provider_readiness.py
@@ -1387,7 +1387,6 @@ from awf.service.provider_readiness_helpers import (  # noqa: E402
     _http_get,
     _http_post_stream,
     _is_cloud_model,
-    _log_redacted_exception,
     _ollama_pull_urls,
     _ollama_tags_urls,
     _ordered_names,
@@ -1398,15 +1397,18 @@ from awf.service.provider_readiness_helpers import (  # noqa: E402
     _probe_ollama,
     _probe_ollama_model,
     _provider_result,
-    _redact,
-    _redact_with_redaction_parts,
     _run_subprocess,
     _runtime_cli_probe_payload,
     _secret_values,
     _security_summary,
     _security_warning,
-    _truncate,
     ensure_ollama_model_available,
+)
+from awf.service.provider_readiness_redaction import (  # noqa: E402
+    _log_redacted_exception,
+    _redact,
+    _redact_with_redaction_parts,
+    _truncate,
 )
 
 __all__ = [
